+++ crates/proto/src/udp/udp_client_stream.rs
@@ -223,13 +223,12 @@ impl Future for UdpResponse {
     type Output = Result<DnsResponse, ProtoError>;
 
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<Self::Output> {
-        dbg!(
-        self.0
+        dbg!(self
+            .0
             .as_mut()
             .poll(cx)
             .map_err(ProtoError::from)
-            .map(|r| r.and_then(|r| r))
-        )
+            .map(|r| r.and_then(|r| r)))
     }
 }