+++ crates/proto/src/xfer/mod.rs
@@ -135,7 +135,7 @@ where
     }
 }
 
-impl<F> Drop for DnsRequestStreamHandle<F> 
+impl<F> Drop for DnsRequestStreamHandle<F>
 where
     F: Future<Output = Result<DnsResponse, ProtoError>> + Send,
 {
@@ -160,8 +160,7 @@ where
 /// The underlying Stream implementation should yield `Some(())` whenever it is ready to send a message,
 ///   NotReady, if it is not ready to send a message, and `Err` or `None` in the case that the stream is
 ///   done, and should be shutdown.
-pub trait DnsRequestSender: Stream<Item = Result<(), ProtoError>> + Send + Unpin + 'static 
-{
+pub trait DnsRequestSender: Stream<Item = Result<(), ProtoError>> + Send + Unpin + 'static {
     /// A future that resolves to a response serial message
     type DnsResponseFuture: Future<Output = Result<DnsResponse, ProtoError>>
         + 'static