+++ crates/resolver/examples/global_resolver.rs
@@ -40,13 +40,13 @@ lazy_static! {
             let mut runtime = tokio::runtime::current_thread::Runtime::new().expect("failed to launch Runtime");
 
             // our platform independent future, result, see next blocks
-            let (resolver, bg) = {
+            let resolver = {
 
                 // To make this independent, if targeting macOS, BSD, Linux, or Windows, we can use the system's configuration:
                 #[cfg(any(unix, windows))]
                 {
                     // use the system resolver configuration
-                    AsyncResolver::from_system_conf().expect("Failed to create AsyncResolver")
+                    AsyncResolver::from_system_conf()
                 }
 
                 // For other operating systems, we can use one of the preconfigured definitions
@@ -62,11 +62,12 @@ lazy_static! {
 
             let &(ref lock, ref cvar) = &*pair2;
             let mut started = lock.lock().unwrap();
+
+            let resolver = runtime.block_on(resolver).expect("failed to create trust-dns-resolver");
+
             *started = Some(resolver);
             cvar.notify_one();
             drop(started);
-
-            runtime.block_on(bg);
         });
 
         // Wait for the thread to start up.
@@ -88,7 +89,7 @@ lazy_static! {
 ///
 /// This looks up the `host` (a `&str` or `String` is good), and combines that with the provided port
 ///   this mimics the lookup functions of `std::net`.
-pub fn resolve<N: IntoName + TryParseIp>(
+pub fn resolve<N: IntoName + TryParseIp + 'static>(
     host: N,
     port: u16,
 ) -> impl Future<Output = io::Result<Vec<SocketAddr>>> {
@@ -113,7 +114,7 @@ pub fn resolve<N: IntoName + TryParseIp>(
     });
 
     // Now return the boxed future
-    Box::new(resolve_future)
+    resolve_future
 }
 
 fn main() {