+++ crates/resolver/src/resolver.rs
@@ -77,9 +77,10 @@ impl Resolver {
         builder.core_threads(1);
 
         let runtime = builder.build()?;
-        let (async_resolver, bg) = AsyncResolver::new(config, options);
-
-        runtime.spawn(bg);
+        let async_resolver = AsyncResolver::new(config, options);
+        let async_resolver = runtime
+            .block_on(async_resolver)
+            .expect("failed to create resolver");
 
         Ok(Resolver {
             runtime: Mutex::new(runtime),
@@ -116,7 +117,9 @@ impl Resolver {
     /// * `name` - name of the record to lookup, if name is not a valid domain name, an error will be returned
     /// * `record_type` - type of record to lookup
     pub fn lookup(&self, name: &str, record_type: RecordType) -> ResolveResult<Lookup> {
-        let lookup = self.async_resolver.lookup(name, record_type);
+        let lookup = self
+            .async_resolver
+            .lookup(name, record_type, Default::default());
         self.runtime.lock()?.block_on(lookup)
     }
 
@@ -127,35 +130,8 @@ impl Resolver {
     /// # Arguments
     ///
     /// * `host` - string hostname, if this is an invalid hostname, an error will be returned.
-    pub fn lookup_ip(&self, host: &str) -> ResolveResult<LookupIp> {
+    pub fn lookup_ip(&self, host: String) -> ResolveResult<LookupIp> {
         let lookup = self.async_resolver.lookup_ip(host);
-        self.runtime.lock()?.block_on(lookup)
-    }
-
-    /// Performs a DNS lookup for an SRV record for the specified service type and protocol at the given name.
-    ///
-    /// This is a convenience method over `lookup_srv`, it combines the service, protocol and name into a single name: `_service._protocol.name`.
-    ///
-    /// # Arguments
-    ///
-    /// * `service` - service to lookup, e.g. ldap or http
-    /// * `protocol` - wire protocol, e.g. udp or tcp
-    /// * `name` - zone or other name at which the service is located.
-    #[deprecated(note = "use lookup_srv instead, this interface is not ideal")]
-    pub fn lookup_service(
-        &self,
-        service: &str,
-        protocol: &str,
-        name: &str,
-    ) -> ResolveResult<lookup::SrvLookup> {
-        #[allow(deprecated)]
-        let lookup = self.async_resolver.lookup_service(service, protocol, name);
-        self.runtime.lock()?.block_on(lookup)
-    }
-
-    /// Lookup an SRV record.
-    pub fn lookup_srv(&self, name: &str) -> ResolveResult<lookup::SrvLookup> {
-        let lookup = self.async_resolver.lookup_srv(name);
         self.runtime.lock()?.block_on(lookup)
     }
 
@@ -165,7 +141,6 @@ impl Resolver {
     lookup_fn!(mx_lookup, lookup::MxLookup);
     lookup_fn!(ns_lookup, lookup::NsLookup);
     lookup_fn!(soa_lookup, lookup::SoaLookup);
-    #[deprecated(note = "use lookup_srv instead, this interface is not ideal")]
     lookup_fn!(srv_lookup, lookup::SrvLookup);
     lookup_fn!(txt_lookup, lookup::TxtLookup);
 }
@@ -189,7 +164,7 @@ mod tests {
     fn test_lookup() {
         let resolver = Resolver::new(ResolverConfig::default(), ResolverOpts::default()).unwrap();
 
-        let response = resolver.lookup_ip("www.example.com.").unwrap();
+        let response = resolver.lookup_ip("www.example.com.".to_string()).unwrap();
         println!("response records: {:?}", response);
 
         assert_eq!(response.iter().count(), 1);
@@ -213,7 +188,7 @@ mod tests {
     fn test_system_lookup() {
         let resolver = Resolver::from_system_conf().unwrap();
 
-        let response = resolver.lookup_ip("www.example.com.").unwrap();
+        let response = resolver.lookup_ip("www.example.com.".to_string()).unwrap();
         println!("response records: {:?}", response);
 
         assert_eq!(response.iter().count(), 1);